emulator: Reparent modules as new entrance; add Touch namespace
Also move RET_ERROR to util namespace.
