+++ src/emulator/CMakeLists.txt
@@ -61,8 +61,8 @@ add_subdirectory(crypto)
 add_subdirectory(ctrl)
 add_subdirectory(dialog)
 add_subdirectory(disasm)
-add_subdirectory(gdbstub)
 add_subdirectory(glutil)
+add_subdirectory(gui)
 add_subdirectory(gxm)
 add_subdirectory(net)
 add_subdirectory(np)
@@ -76,13 +76,17 @@ add_subdirectory(nids)
 add_subdirectory(renderer)
 add_subdirectory(rtc)
 add_subdirectory(shader)
-add_subdirectory(util)
 add_subdirectory(threads)
-add_subdirectory(gui)
+add_subdirectory(touch)
+add_subdirectory(util)
+
+if(USE_GDBSTUB)
+	add_subdirectory(gdbstub)
+endif()
 
 add_executable(emulator MACOSX_BUNDLE main.cpp performance.cpp)
 
-target_link_libraries(emulator PRIVATE app discord-rpc gui host sdl2 shader util)
+target_link_libraries(emulator PRIVATE modules)
 if(LINUX)
 	target_link_libraries(emulator PRIVATE stdc++fs)
 endif()
