+++ src/emulator/app/CMakeLists.txt
@@ -3,15 +3,17 @@ add_library(
 	STATIC
 	include/app/app_functions.h
 	include/app/discord.h
+	include/app/load_self.h
 	include/app/screen_render.h
 	src/app_init.cpp
 	src/app_private.h
 	src/app_sfo.cpp
 	src/app.cpp
 	src/discord.cpp
+	src/load_self.cpp
 	src/screen_render.cpp
 )
 
 target_include_directories(app PUBLIC include)
-target_link_libraries(app PUBLIC glutil host util)
-target_link_libraries(app PRIVATE ${Boost_LIBRARIES} audio discord-rpc gui io kernel microprofile rtc sdl2 spdlog yaml-cpp)
+target_link_libraries(app PUBLIC host vita-headers)
+target_link_libraries(app PRIVATE elfio vita-toolchain)
