+++ src/emulator/app/include/app/load_self.h
