+++ src/emulator/app/src/load_self.cpp
@@ -15,8 +15,9 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
+#include <app/load_self.h>
+
 #include <config/config.h>
-#include <kernel/load_self.h>
 #include <kernel/relocation.h>
 #include <kernel/state.h>
 #include <kernel/types.h>
