+++ src/emulator/io/CMakeLists.txt
@@ -13,4 +13,3 @@ add_library(
 
 target_include_directories(io PUBLIC include)
 target_link_libraries(io PUBLIC rtc miniz vita-headers dirent mem)
-target_link_libraries(io PRIVATE util)
