+++ src/emulator/module/CMakeLists.txt
@@ -5,16 +5,18 @@ add_library(
 	include/module/bridge.h
 	include/module/bridge_types.h
 	include/module/lay_out_args.h
+	include/module/load_module.h
 	include/module/module.h
 	include/module/read_arg.h
 	include/module/write_return_value.h
+	src/load_module.cpp
 	src/module.cpp
 	src/read_arg.cpp
 	src/write_return_value.cpp
 )
 
 target_include_directories(module PUBLIC include)
-target_link_libraries(module PUBLIC kernel host microprofile rpcs3 dlmalloc printf)
+target_link_libraries(module PUBLIC app dlmalloc printf)
 
 add_executable(
 	module-tests
