+++ src/emulator/module/src/module.cpp
@@ -54,18 +54,3 @@ int stubbed_impl(const char *name, const char *info) {
 
     return 0;
 }
-
-int ret_error_impl(const char *name, const char *error_str, std::uint32_t error_val) {
-    bool inserted = false;
-
-    {
-        const std::lock_guard<std::mutex> lock(mutex);
-        inserted = logged.insert(name).second;
-    }
-
-    if (inserted) {
-        LOG_ERROR("{} returned {} ({})", name, error_str, log_hex(error_val));
-    }
-
-    return error_val;
-}
