+++ src/emulator/util/CMakeLists.txt
@@ -17,4 +17,4 @@ add_library(
 )
 
 target_include_directories(util PUBLIC include)
-target_link_libraries(util PUBLIC spdlog)
+target_link_libraries(util PUBLIC ${Boost_LIBRARIES} spdlog)
